css updates for vertical nav and new JS feature
added in styles for vertical nav, also the javascript for makinh the generate button lead straight to the answer on the next frame so the user doesn't have to scroll. this allows the user to be led the their generated prompt as soon as they click generate. they dont have to scroll by themselves